// src/model/GenreId.tsx
export enum GenreId {
    action = 28,
    adventure = 12,
    animation = 16,
    comedy = 35,
    crime = 80,
    documentary = 99,
    drama = 18,
    family = 10751,
    fantasy = 14,
    history = 36,
    horror = 27,
    music = 10402,
    mystery = 9648,
    romance = 10749,
    scienceFiction = 878,
    tvMovie = 10770,
    thriller = 53,
    war = 10752,
    western = 37
}